refactor: prepare joining data and maps
