+++ src/modes/lab/CountyRelated.svelte
@@ -1,7 +1,7 @@
 <script>
   import Vega from '../../components/vega/Vega.svelte';
   import { addMissing, fetchTimeSlice, averageByDate } from '../../data';
-  import { getInfoByName, nationInfo, getStateOfCounty } from '../../maps/infos';
+  import { getInfoByName, nationInfo, getStateOfCounty } from '../../maps';
   import getRelatedCounties from '../../maps/related';
   import { currentDateObject } from '../../stores';
   import { defaultRegionOnStartup, DEFAULT_SURVEY_SENSOR, sensorMap } from '../../stores/constants';
